Crate hotg_rune_compiler

The Rune Compiler

A compiler that compiles your data processing pipeline into a portable WebAssembly binary.

Architecture

The Rune compiler is base around Salsa, a library for incremental computation. This lets us phrase the compilation process as a series of queries (essentially, pure functions) which can be aggressively cached based on dependency analysis.

These series of queries are broken up into a couple submodules,

parse - Parse a Runefile written in the YAML format

codegen - Generate the final Rune binary

Re-exports

pub use crate::im::Text;

Modules

asset_loader

codegen

im

Immutable collections.

parse

The YAML frontend for the Rune compiler.

type_check

The Rune compiler’s type checker.

Structs

BuildConfig

EnvironmentStorage

Representative struct for the query group.

FeatureFlags

Flags used by the Rune compiler to enable features.

Traits

Environment

The build environment.

Type Definitions

Error