Module hotg_rune_compiler::codegen

The code generation phase.

This takes the parsed and analysed Rune and generates all the necessary files to make a Rust project.

Structs

CapabilitySummary

CustomSection

A WebAssembly custom section to be embedded in the Rune.

File

A file that will be written to the Rune’s build directory.

ModelSummary

OutputSummary

ProcBlockSummary

RuneGraph

A summary of the Rune pipeline that will be embedded in the Rune.

RuneSummary

RuneVersion

TensorId

Constants

GRAPH_CUSTOM_SECTION

RESOURCE_CUSTOM_SECTION

VERSION_CUSTOM_SECTION

Functions

phase