Module hotg_rune_compiler::hooks

Callbacks that allow users to hook into the build process.

Enums

Continuation

How to proceed after calling a Hooks method.

Traits

AfterCodegenContext

Context passed to the Hooks::after_codegen() method.

AfterCompileContext

Context passed to the Hooks::after_compile() method.

AfterLoweringContext

Context passed to the Hooks::after_lowering() method.

AfterParseContext

Context passed to the Hooks::after_parse() method.

AfterTypeCheckingContext

Context passed to the Hooks::after_type_checking() method.

Context

Basic contextual information passed to all Hooks.

Hooks

Callbacks that are fired at different points in the compilation process.