Module hotg_rune_compiler::lowering

The lowering phase.

Structs

Inputs

The list of Tensors that may be the inputs to a PipelineNode.

Mimetype

Model

A ML model.

ModelData

Name

An identifier used to refer to an item in a Runefile.

NameTable

A lookup table mapping Name components back to their Entity.

Outputs

The list of Tensors that may be the output from a PipelineNode.

PipelineNode

A tag component indicating this Entity is part of the Rune’s pipeline.

ProcBlock

Resource

ResourceData

Sink

An output.

Source

Something which can generate data.

Tensor

The Shape a tensor may take.

Enums

ModelFile

Where to load a model from.

ResourceOrString

ResourceSource

SinkKind

The kind of output.

SourceKind

Where should a Source pull its data from?

Functions

phase