Trait hotg_rune_compiler::parse::Frontend

pub trait Frontend: Database + HasQueryGroup<FrontendStorage> + Environment + AssetLoader {
    fn src(&self) -> Text;
    fn set_src(&mut self, value__: Text);
    fn set_src_with_durability(
        &mut self, 
        value__: Text, 
        durability__: Durability
    );
    fn parse(&self) -> Result<Arc<DocumentV1>, Error>;
    fn parse_runefile(&self, key0: Text) -> Result<Arc<Document>, ParseFailed>;
    fn resource_value(&self, key0: Text) -> Result<Vector<u8>, Error>;
    fn resource_values(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>;
    fn proc_block(&self, key0: Text) -> Result<Vector<u8>, Error>;
    fn proc_blocks(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>;
    fn model_file(&self, key0: Text) -> Result<Vector<u8>, Error>;
    fn model_files(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>;
}

The Rune compiler’s YAML frontend.

Examples

You will typically use the Frontend by first setting the Runefile.yml source code (Frontend::set_src()) and then parse the document with Frontend::parse().

use hotg_rune_compiler::{
    parse::{Frontend, FrontendStorage},
    asset_loader::{AssetLoader, ReadError},
    parse::Path, EnvironmentStorage, im::Vector,
};
use uriparse::URI;

// First, you need to create a database which can hold Salsa's state

#[derive(Default)]
#[salsa::database(FrontendStorage, EnvironmentStorage)]
struct Database {
    storage: salsa::Storage<Self>,
}

impl salsa::Database for Database {}

// The parsing process requires you to load proc-blocks and other assets.
// You can satisfy these dependencies by implementing the corresponding
// traits.

impl AssetLoader for Database {
    fn read(&self, path: &URI<'_>) -> Result<Vector<u8>, ReadError> {
        todo!();
    }
}

// Let's hard-code a YAML document to parse
let runefile = r#"
    version: 1
    image: runicos/base
    pipeline:
      input:
        capability: RAND
        outputs:
        - type: f32
          dimensions: [1]
"#;

// Instantiate the database
let mut db = Database::default();

// Set the source
db.set_src(runefile.into());

// And parse it
let doc = db.parse().unwrap();

Required Methods

fn src(&self) -> Text

The YAML document being parsed.

fn set_src(&mut self, value__: Text)

Set the value of the src input.

See src for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_src_with_durability(&mut self, value__: Text, durability__: Durability)

Set the value of the src input and promise that its value will never change again.

See src for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn parse(&self) -> Result<Arc<DocumentV1>, Error>

Parse the Frontend::src() into a DocumentV1.

fn parse_runefile(&self, key0: Text) -> Result<Arc<Document>, ParseFailed>

A low-level query for parsing a YAML file.

fn resource_value(&self, key0: Text) -> Result<Vector<u8>, Error>

Get a resource’s value by either returning the value as-is (for inline resources) or reading the file from the filesystem.

fn resource_values(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>

fn proc_block(&self, key0: Text) -> Result<Vector<u8>, Error>

fn proc_blocks(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>

Get the binary data for each proc-block in this Rune, ignoring any which may have failed to load.

fn model_file(&self, key0: Text) -> Result<Vector<u8>, Error>

fn model_files(&self) -> Result<OrdMap<Text, Vector<u8>>, Error>