Trait hotg_rune_compiler::Environment

pub trait Environment: Database + HasQueryGroup<EnvironmentStorage> {
    fn config(&self) -> BuildConfig;
    fn set_config(&mut self, value__: BuildConfig);
    fn set_config_with_durability(
        &mut self, 
        value__: BuildConfig, 
        durability__: Durability
    );
}

The build environment.

Required Methods

fn config(&self) -> BuildConfig

fn set_config(&mut self, value__: BuildConfig)

Set the value of the config input.

See config for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.

fn set_config_with_durability(
&mut self,
value__: BuildConfig,
durability__: Durability
)

Set the value of the config input and promise that its value will never change again.

See config for details.

Note: Setting values will trigger cancellation of any ongoing queries; this method blocks until those queries have been cancelled.