Crate hotg_rune_compiler

The Rune compiler.

Phases

The compilation process is split into several phases which build on each other, with user-injectable hooks called after each phase finishes.

The phases are:

parse

lowering

type_check

codegen

Stability

This crate contains the internal types used by the Rune compiler so they can be used externally. While this can give you a lot of flexibility and let you extract a lot of information about a Rune, the compiler is a continually evolving codebase.

This API should be considered unstable and subject to change.

Modules

codegen

The code generation phase.

compile

hooks

Callbacks that allow users to hook into the build process.

lowering

The lowering phase.

parse

The parsing phase.

serialize

type_check

The type checking phase.

Structs

BuildContext

Inputs used during the compilation process.

Diagnostics

A collection of Diagnostics.

FeatureFlags

Feature flags and other knobs that can be used during development.

Phase

A group of operations which make up a single “phase” in the build process.

Enums

Verbosity

Functions

build

Execute the rune build process.

build_with_hooks

Execute the rune build process, passing in custom Hooks which will be fired after each phase.

rust_toolchain

Get a copy of the rust-toolchain.toml file used by the Rune project itself.