Module hotg_rune_compiler::parse

The parsing phase.

This is a simple phase which just calls Document::parse() and stores the resulting DocumentV1 in the global legion::Resources.

Structs

Argument

A newtype around ResourceOrString which is used in each stage’s args dictionary.

CapabilityStage

A stage which reads inputs from the runtime.

DocumentV1

Version 1 of the Runefile.yml format.

Image

The image a Rune is based on.

Input

The name of a tensor.

ModelStage

A ML model which will be executed by the runtime.

OutStage

A stage which passes outputs back to the runtime.

Path

A specification for finding a dependency.

PathParseError

ProcBlockStage

A stage which executes a procedural block.

ResourceDeclaration

The declaration for a resource, typically something like a wordlist or environment variable.

ResourceName

A reference to some ResourceDeclaration. It typically looks like $RESOURCE_NAME.

Type

The element type and dimensions for a particular tensor.

Enums

Document

The top level Runefile type.

ResourceOrString

Something that could be either a reference to a resource ($resource) or a plain string (./path).

ResourceType

How the resource should be treated inside the Rune.

Stage

A stage in the Rune’s pipeline.

Functions

phase